requirements.txt
streamlit==1.54.0
numpy==2.4.2
Pillow==12.1.1